Create App.js as a separate module
